project filter changed
